refactor : Cord refactoring to simple
