+++ src/main/java/com/example/daon/accounting/expenseProof/service/ExpenseProofService.java
@@ -8,6 +8,7 @@ import com.example.daon.accounting.expenseProof.repository.ExpenseProofRepositor
 import com.example.daon.customer.model.CustomerEntity;
 import com.example.daon.customer.repository.CustomerRepository;
 import com.example.daon.global.exception.ResourceInUseException;
+import com.example.daon.global.service.ConvertResponseService;
 import com.example.daon.global.service.GlobalService;
 import com.example.daon.receipts.model.FromCategory;
 import com.example.daon.receipts.model.ReceiptCategory;
@@ -17,7 +18,6 @@ import jakarta.persistence.criteria.Predicate;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -32,6 +32,7 @@ public class ExpenseProofService {
     private final CustomerRepository customerRepository;
     private final ReceiptRepository receiptRepository;
     private final CategorySelectionService categorySelectionService;
+    private final ConvertResponseService convertResponseService;
     private final GlobalService globalService;
 
     //지출증빙
@@ -79,7 +80,7 @@ public class ExpenseProofService {
             // 동적 조건을 조합하여 반환
             return criteriaBuilder.and(predicates.toArray(new Predicate[0]));
         });
-        return expenseProofEntities.stream().map(globalService::convertToExpenseProofResponse).collect(Collectors.toList());
+        return expenseProofEntities.stream().map(convertResponseService::convertToExpenseProofResponse).collect(Collectors.toList());
     }
 
     public void paidExpenseProof(ExpenseProofRequest expenseProofRequest) {
@@ -101,7 +102,7 @@ public class ExpenseProofService {
                     expenseProofEntity.getMemo(),
                     FromCategory.SALES));
             expenseProofEntity.setReceiptId(receipt.getReceiptId());
-            expenseProofEntity.setPaidDate(LocalDate.now());
+            expenseProofEntity.setPaidDate(expenseProofRequest.getPaidDate());
             globalService.updateDailyTotal(receipt.getTotalPrice(), receipt.getCategory(), receipt.getTimeStamp());
         } else {
             ReceiptEntity receipt = receiptRepository.findById(expenseProofEntity.getReceiptId()).orElseThrow(() -> new RuntimeException("전표가 존재하지 않습니다."));
