+++ src/main/java/com/example/daon/accounting/salesVAT/service/SalesVATService.java
@@ -8,6 +8,7 @@ import com.example.daon.accounting.salesVAT.repository.SalesVATRepository;
 import com.example.daon.customer.model.CustomerEntity;
 import com.example.daon.customer.repository.CustomerRepository;
 import com.example.daon.global.exception.ResourceInUseException;
+import com.example.daon.global.service.ConvertResponseService;
 import com.example.daon.global.service.GlobalService;
 import com.example.daon.receipts.model.FromCategory;
 import com.example.daon.receipts.model.ReceiptCategory;
@@ -17,7 +18,6 @@ import jakarta.persistence.criteria.Predicate;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -31,6 +31,7 @@ public class SalesVATService {
     private final CustomerRepository customerRepository;
     private final ReceiptRepository receiptRepository;
     private final CategorySelectionService categorySelectionService;
+    private final ConvertResponseService convertResponseService;
     private final GlobalService globalService;
 
     //매출부가세
@@ -79,7 +80,7 @@ public class SalesVATService {
             return criteriaBuilder.and(predicates.toArray(new Predicate[0]));
         });
         System.out.println(salesVATEntities.size());
-        return salesVATEntities.stream().map(globalService::convertToSalesVATResponse).collect(Collectors.toList());
+        return salesVATEntities.stream().map(convertResponseService::convertToSalesVATResponse).collect(Collectors.toList());
     }
 
     public void salesVATPaid(SalesVATRequest salesVATRequest) {
@@ -101,7 +102,7 @@ public class SalesVATService {
                     salesVATEntity.getMemo(),
                     FromCategory.SALES));
             salesVATEntity.setReceiptId(receipt.getReceiptId());
-            salesVATEntity.setPaidDate(LocalDate.now());
+            salesVATEntity.setPaidDate(salesVATRequest.getPaidDate());
             salesVATRequest.setReceiptId(salesVATEntity.getReceiptId());
             globalService.updateDailyTotal(receipt.getTotalPrice(), receipt.getCategory(), receipt.getTimeStamp());
         } else {
